# api-specs/api_spec_PGS.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/
    description: Generated server url
paths:
  /bpay-payment-response/{id}:
    get:
      tags:
        - payment-transactions-controller
      operationId: getBPayPaymentResponse
      parameters:
        - in: path
          name: id
          schema:
            type: number
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BPayPaymentResponseEntity'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Response not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Generic Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "504":
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /request-payments/bancomatpay:
    put:
      tags:
        - payment-transactions-controller
      operationId: updateTransaction
      parameters:
        - name: X-Correlation-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthMessage'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ACKMessage'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized request correlationId already processes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Unknown correlationId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "424":
          description: Error in transaction patch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Generic Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "504":
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - payment-transactions-controller
      operationId: requestPaymentToBancomatPay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BPayPaymentRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BPayOutcomeResponse'
        "400":
          description: Bad Request mandatory parameters missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: transactionId already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Generic Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "504":
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /request-refunds/bancomatpay:
    post:
      tags:
        - payment-transactions-controller
      operationId: refundTransaction
      parameters:
        - name: X-Correlation-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BPayRefundRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BPayOutcomeResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Generic Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "504":
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /request-payments/postepay/{requestId}:
    get:
      summary: PGS webview polling call for PostePay authorization
      tags:
        - payment-transactions-controller
      operationId: webview-polling
      parameters:
        - in: path
          name: requestId
          schema:
            type: string
            format: uuid
          required: true
          description: PGS-generated GUID of the request to retrieve
          example: 77e1c83b-7bb0-437b-bc50-a7a58e5660ac
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PollingResponseEntity'
        "400":
          description: Bad request - missing mandatory parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollingResponseEntity'
        "401":
          description: transactionId already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollingResponseEntity'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollingResponseEntity'
        "504":
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollingResponseEntity'
    delete:
      tags:
        - payment-transactions-controller
      summary: refund PostePay requests
      operationId: refund-request
      parameters:
        - in: path
          name: requestId
          schema:
            type: string
            format: uuid
          required: true
          description: PGS-generated GUID of the request to retrieve
          example: 77e1c83b-7bb0-437b-bc50-a7a58e5660ac
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayRefundResponse'
        "404":
          description: Request doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayRefundResponse'
        "500":
          description: Generic Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayRefundResponse'
        "502":
          description: Gateway Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayRefundResponse'
        "504":
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayRefundResponse'
  /request-payments/postepay:
    post:
      summary: payment authorization request to PostePay
      tags:
        - payment-transactions-controller
      operationId: auth-request
      parameters:
        - in: header
          name: X-Client-ID
          description: channel origin (APP/Web)
          example: 77e1c83b-7bb0-437b-bc50-a7a58e5660ac
          schema:
            type: string
            format: uuid
          required: true
        - in: header
          name: MDC-Fields
          description: MDC information
          example: 97g10t83x7bb0437bbc50sdf58e970gt
          schema:
            type: string
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostePayAuthRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PostePayAuthResponseEntity'
        "400":
          description: Bad request - missing mandatory parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayAuthResponseEntity'
        "401":
          description: transactionId already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayAuthResponseEntity'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayAuthResponseEntity'
        "504":
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostePayAuthResponseEntity'
    put:
      tags:
        - payment-transactions-controller
      summary: authorization outcome response from PostePay
      operationId: auth-response
      parameters:
        - name: X-Correlation-ID
          in: header
          required: true
          description: PostePay correlation ID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthMessage'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ACKMessage'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized - idPostePay already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Unknown idPostePay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Generic Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "504":
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    AuthMessage:
      required:
        - auth_outcome
      type: object
      properties:
        auth_outcome:
          type: string
          enum:
            - OK
            - KO
        auth_code:
          type: string
    ACKMessage:
      required:
        - outcome
      type: object
      properties:
        outcome:
          type: string
          enum:
            - OK
            - KO
    BPayPaymentRequest:
      required:
        - amount
        - encryptedTelephoneNumber
        - idPagoPa
        - idPsp
      type: object
      properties:
        idPsp:
          type: string
        idPagoPa:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        subject:
          type: string
        encryptedTelephoneNumber:
          type: string
        language:
          type: string
    BPayRefundRequest:
      required:
        - idPagoPa
      type: object
      properties:
        idPagoPa:
          type: integer
          format: int64
        refundAttempt:
          type: number
          format: integer
        subject:
          type: string
        language:
          type: string
    BPayOutcomeResponse:
      type: object
      properties:
        outcome:
          type: boolean
    PollingResponseEntity:
      type: object
      required:
        - channel
        - urlRedirect
        - logoResourcePath
        - clientResponseUrl
        - authOutcome
        - error
      properties:
        channel:
          type: string
          description: request payment channel (APP or WEB)
          example: APP
        urlRedirect:
          type: string
          description: redirect URL generated by PGS logic
          example: "https://.../payment-transactions-gateway/v1/webview/authRequest?requestId=e2f518a9-9de4-4a27-afb0-5303e6eefcbf"
        logoResourcePath:
          type: string
          description: PostePay logo resource path
          example: "payment-gateway/assets/img/postepay/postepay.png"
        clientResponseUrl:
          type: string
          description: redirect URL for authorization OK, empty for KO case or when PUT request has not been called yet
          example: "https://..."
        authOutcome:
          type: string
          description: authorization outcome
          example: OK
        error:
          type: string
          description: "error description for 400/500 http error codes"
          example: ""
    PostePayAuthRequest:
      required:
        - grandTotal
        - idTransaction
      type: object
      properties:
        grandTotal:
          type: number
          format: integer
          description: amount + fee as euro cents
          example: 2350
        idTransaction:
          type: integer
          format: int64
          description: transaction id on Payment Manager
          example: 123456
        description:
          type: string
          description: payment object description
          example: pagamento bollo auto
        paymentChannel:
          type: string
          description: request payment channel (APP or WEB)
          example: APP
        name:
          type: string
          description: debtor's name and surname
          example: Mario Rossi
        email_notice:
          type: string
          description: debtor's email address
          example: mario.rossi@gmail.com
    PostePayAuthResponseEntity:
      type: object
      required:
        - requestId
        - channel
        - urlRedirect
        - error
      properties:
        requestId:
          type: string
          description: payment request Id
        channel:
          type: string
          description: request payment channel (APP or WEB)
          example: APP
        urlRedirect:
          type: string
          description: redirect URL generated by PGS logic
          example: "https://api.dev.platform.pagopa.it/payment-transactions-gateway/v1/webview/authRequest/e8d24dd5-14e0-4802-9f23-ddf2c198a185"
        error:
          type: string
          description: error description
          example: null #valued when a different response from 200 is obtained
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int64
        message:
          type: string
      required:
        - code
        - message
    PostePayRefundResponse:
      type: object
      properties:
        requestId:
          type: string
        paymentId:
          type: string
        refundOutcome:
          type: string
        error:
          type: string
      required:
        - transactionId
        - paymentId
        - refundOutcome
        - error